use 20-digit fids in md5 table
git-svn-id: http://dev.oxf.freelogy.org/svn/polony/polony-tools/trunk@85 1b2bb62c-e52f-0410-8f12-f050cf38ac25

+++ www/inc-mogilefs.php
@@ -19,7 +19,7 @@ echo mysql_error();
 
 mysql_query("create table if not exists md5
 (
- fid int(10) unsigned not null,
+ fid bigint(20) unsigned not null,
  md5 char(32),
  primary key (fid)
 ) engine=innodb");
